+++ tests/test_github_create_pr.py
@@ -314,7 +314,7 @@ class TestGitHubCreatePR(BaseTest):
         check_out("allow-ownership-link")
         assert_success(
             ["github", "create-pr", "--title=PR title set explicitly"],
-            f"""
+            """
             Push allow-ownership-link to origin? (y, N, q)
 
               master
@@ -343,7 +343,7 @@ class TestGitHubCreatePR(BaseTest):
             Setting milestone of PR #7 to 42... OK
             Adding github_user as assignee to PR #7... OK
             Adding invalid-user as reviewer to PR #7...
-            Warn: there are some invalid reviewers (non-collaborators) in .git{os.path.sep}info{os.path.sep}reviewers file.
+            Warn: there are some invalid reviewers (non-collaborators) in .git/info/reviewers file.
             Skipped adding reviewers to the pull request.
             OK
             """