+++ RELEASE_NOTES.md
@@ -4,6 +4,7 @@
 
 - fixed: suggested `update` command in deprecation message for `fork-point --override-to=...` and `--override-to-inferred`
 - fixed: subtle bugs related to relative main/git directory paths when switching worktrees in `traverse`
+- changed: whenever git-machete prints a path on Windows, the path is now POSIX-compatible (with `/`)
 
 ## New in git-machete 3.39.0